backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
openpyxl==3.1.5
python-decouple==3.8
